html/dom/elements/global-attributes/dataset-get.html

<!DOCTYPE html>
<html>
<head>
<title>Dataset - Get</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<h1>Dataset - Get</h1>
<div id="log"></div>
<script>
function testGet(attr, expected, name) {
  test(function() {
    var d = document.createElement("div");
    d.setAttribute(attr, "value");
    assert_equals(d.dataset[expected], "value");
  }, name);
}

testGet('data-foo', 'foo',
  "Getting element.dataset['foo'] should return the value of element.getAttribute('data-foo')'");
testGet('data-foo-bar', 'fooBar',
  "Getting element.dataset['fooBar'] should return the value of element.getAttribute('data-foo-bar')'");
testGet('data--', '-',
  "Getting element.dataset['-'] should return the value of element.getAttribute('data--')'");
testGet('data--foo', 'Foo',
  "Getting element.dataset['Foo'] should return the value of element.getAttribute('data--foo')'");
testGet('data---foo', '-Foo',
  "Getting element.dataset['-Foo'] should return the value of element.getAttribute('data---foo')'");
testGet('data-Foo', 'foo',
  "Getting element.dataset['foo'] should return the value of element.getAttribute('data-Foo')'");
testGet('data-', '',
  "Getting element.dataset[''] should return the value of element.getAttribute('data-')'");
testGet('data-\xE0', '\xE0',
  "Getting element.dataset['\xE0'] should return the value of element.getAttribute('data-\xE0')'");
testGet('data-to-string', 'toString',
  "Getting element.dataset['toString'] should return the value of element.getAttribute('data-to-string')'");

function size(dataset)
{
  var count = 0;
  for (var item in dataset) {
    count++;
  }
  return count;
}

test(function() {
  var d = document.createElement("div");
  d.setAttribute('dataFoo', "value");
  assert_equals(size(d.dataset), 0);
}, "Tests that an attribute named dataFoo does not make an entry in the dataset DOMStringMap.");
</script>
</body>
</html>
